Created separate window to display debug information about all objects on the space
