+++ SpaceIntruders/View/MainWindow.xaml
@@ -12,24 +12,6 @@
     </Window.DataContext>
 
     <Window.Resources>
-        <DataTemplate x:Key = "shipsTemplate">
-            <StackPanel Orientation = "Horizontal">
-
-                <TextBox Text = "{Binding Path = Name, Mode = OneWay}" 
-               Width = "100" Margin = "3 5 3 5"/>
-                <TextBox Text = "{Binding Path = X, Mode = OneWay}" 
-               Width = "50" Margin = "0 5 3 5"/>
-                <TextBox Text = "{Binding Path = Y, Mode = OneWay}" 
-               Width = "50" Margin = "0 5 3 5"/>
-                <TextBox Text = "{Binding Path = Width, Mode = OneWay}" 
-               Width = "50" Margin = "0 5 3 5"/>
-                <TextBox Text = "{Binding Path = Height, Mode = OneWay}" 
-               Width = "50" Margin = "0 5 3 5"/>
-
-            </StackPanel>
-
-        </DataTemplate>
-
         <DataTemplate x:Key = "shipsImageTemplate">
             <Image x:Name="image" HorizontalAlignment="Left" Margin="{Binding Path = Margin}" VerticalAlignment="Top" Source="{Binding Path = ImageSouce}" Width="100" Height="100"/>
         </DataTemplate>
@@ -37,32 +19,24 @@
 
     <Grid>
         <StackPanel>
-            <ListView>
+            <ItemsControl ItemsSource="{Binding EnvironmentObjects}" Margin="10,10,10,10" Width="{Binding CosmoSpace.Width}" Height="{Binding CosmoSpace.Heigth}">
 
-                <ListView ItemsSource="{Binding EnvironmentObjects}" 
-                     ItemTemplate="{StaticResource shipsTemplate}"
-                     Margin="10,10,10,10" />
+                <ItemsControl.Background>
+                    <SolidColorBrush x:Name="BorderBrush" Color="Black"/>
+                </ItemsControl.Background>
 
-                <ItemsControl ItemsSource="{Binding EnvironmentObjects}" Margin="10,10,10,10" Width="{Binding CosmoSpace.Width}" Height="{Binding CosmoSpace.Heigth}">
+                <ItemsControl.ItemsPanel>
+                    <ItemsPanelTemplate>
+                        <Canvas/>
+                    </ItemsPanelTemplate>
+                </ItemsControl.ItemsPanel>
 
-                    <ItemsControl.Background>
-                        <SolidColorBrush x:Name="BorderBrush" Color="Black"/>
-                    </ItemsControl.Background>
-
-                    <ItemsControl.ItemsPanel>
-                        <ItemsPanelTemplate>
-                            <Canvas/>
-                        </ItemsPanelTemplate>
-                    </ItemsControl.ItemsPanel>
-
-                    <ItemsControl.ItemTemplate>
-                        <DataTemplate >
-                            <Image HorizontalAlignment="Left" Margin="{Binding Path = Margin}" VerticalAlignment="Top" Source="{Binding Path = ImageSouce}" Width="{Binding Path = Width}" Height="{Binding Path = Height}"/>
-                        </DataTemplate>
-                    </ItemsControl.ItemTemplate>
-                </ItemsControl>
-
-            </ListView>
+                <ItemsControl.ItemTemplate>
+                    <DataTemplate >
+                        <Image HorizontalAlignment="Left" Margin="{Binding Path = Margin}" VerticalAlignment="Top" Source="{Binding Path = ImageSouce}" Width="{Binding Path = Width}" Height="{Binding Path = Height}"/>
+                    </DataTemplate>
+                </ItemsControl.ItemTemplate>
+            </ItemsControl>
 
             <StackPanel Orientation="Horizontal" HorizontalAlignment="Center">
                 <RepeatButton Command ="{Binding MoveLeft}" Width="50" Height="50" Margin="0,0,0,0" Content="&lt;"></RepeatButton>
