+++ SpaceIntruders/View/ObjectInfoWindow.xaml
@@ -1,0 +1,35 @@
+﻿<Window x:Class="SpaceIntruders.View.ObjectInfoWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:SpaceIntruders.View"
+        mc:Ignorable="d"
+        Title="ObjectInfoWindow" Height="300" Width="552.733">
+    <Window.Resources>
+        <DataTemplate x:Key = "shipsTemplate">
+            <StackPanel Orientation = "Horizontal">
+
+                <TextBox Text = "{Binding Path = Name, Mode = OneWay}" 
+               Width = "100" Margin = "3 5 3 5"/>
+                <TextBox Text = "{Binding Path = X, Mode = OneWay}" 
+               Width = "50" Margin = "0 5 3 5"/>
+                <TextBox Text = "{Binding Path = Y, Mode = OneWay}" 
+               Width = "50" Margin = "0 5 3 5"/>
+                <TextBox Text = "{Binding Path = Width, Mode = OneWay}" 
+               Width = "50" Margin = "0 5 3 5"/>
+                <TextBox Text = "{Binding Path = Height, Mode = OneWay}" 
+               Width = "50" Margin = "0 5 3 5"/>
+            </StackPanel>
+
+        </DataTemplate>
+
+    </Window.Resources>
+
+    <Grid>
+        <ListView ItemsSource="{Binding EnvironmentObjects}" 
+                     ItemTemplate="{StaticResource shipsTemplate}"
+                     Margin="10,10,10,10" />
+    </Grid>
+    
+</Window>
